Apply feedback and refactor folder structure
