+++ package.json
@@ -10,7 +10,6 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.22.0",
     "react-scripts": "^5.0.1",
-    "sass": "^1.71.0",
     "styled-components": "^6.1.8",
     "web-vitals": "^2.1.4"
   },
