+++ src/components/Footer/Footer.jsx
@@ -1,4 +1,4 @@
-import "./styles/Footer.css";
+import "./Footer.css";
 
 export default function Footer() {
   return (
