+++ src/components/index.js
@@ -1,0 +1,21 @@
+import Card from "./Card/Card";
+import Cards from "./Cards/Cards";
+import FolderPageHeader from "./FolderPageHeader/FolderPageHeader";
+import FoldersList from "./FoldersList/FoldersList";
+import FolderToolbar from "./FolderToolbar/FolderToolbar";
+import Footer from "./Footer/Footer";
+import Navbar from "./Navbar/Navbar";
+import Searchbar from "./Searchbar/Searchbar";
+import SharedPageHeader from "./SharedPageHeader/SharedPageHeader";
+
+export {
+  Card,
+  Cards,
+  FolderPageHeader,
+  FolderToolbar,
+  FoldersList,
+  Footer,
+  Navbar,
+  Searchbar,
+  SharedPageHeader,
+};
