+++ src/components/styled/LinkAddCTA.jsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+import CTA from "./CTA";
+
+const LinkAddCTA = styled(CTA)`
+  width: 80px;
+  padding: 10px 16px;
+  font-size: 14px;
+`;
+
+export default LinkAddCTA;
