+++ src/pages/SharedPage/style.css
